requirements.txt
flask
gunicorn
flask-cors
pillow
pillow-heif
